+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Migrate LLM Evaluation logic to scripts and add tests
+- Fix /threads/messages endpoint.
 
 ## [0.3.3] - 30.10.2024